@@ -1,9 +1,9 @@
 package com.briot.tmmlmss.implementor.ui.main
 
-import android.arch.lifecycle.ViewModelProviders
+import androidx.lifecycle.ViewModelProviders
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+import androidx.appcompat.app.AppCompatActivity
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -13,7 +13,7 @@ import com.briot.tmmlmss.implementor.repository.local.PrefRepository
 import kotlinx.android.synthetic.main.produce_details_fragment.*
 
 
-class ProduceDetailsFragment : Fragment() {
+class ProduceDetailsFragment : androidx.fragment.app.Fragment() {
 
     companion object {
         fun newInstance() = ProduceDetailsFragment()
