@@ -1,8 +1,8 @@
 package com.briot.tmmlmss.implementor.ui.main
 
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.ViewModel;
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel;
 import android.util.Log
 import com.briot.tmmlmss.implementor.repository.remote.Asset
 import com.briot.tmmlmss.implementor.repository.remote.RemoteRepository
