@@ -1,6 +1,6 @@
 package com.briot.tmmlmss.implementor.ui.main
 
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.ViewModel
 import com.briot.tmmlmss.implementor.SQLConnection
 
 class MainViewModel : ViewModel() {
